Re-Organization for CSS optimization
To be able to include icon-16|32.css in another file that is above
those CSS. This hard coding of the relative path w.r.t. to another CSS
is not very clean but it does what I want. Example:

  .icon-img-16 { bg-img: url(ico/fatcow/16x16/img.png) !important; }

instead of the actual cleaner

  .icon-img-16 { bg-img: url(16x16/img.png) !important; }

That means that the icon-16|32.css are required to be included an
another CSS ../../min.css where the relative path w.r.t. to min.css has
to be ico/fatcow/16x16 or ico/fatcow/32x32.